+++ AddOns/BetterBags/data/refresh.lua
@@ -61,10 +61,6 @@ function refresh:StartUpdate()
     elseif event.eventName == 'EQUIPMENT_SETS_CHANGED' then
       updateBackpack = true
       updateBank = true
-    elseif event.eventName == 'PLAYERBANKSLOTS_CHANGED' then
-      updateBank = true
-    elseif event.eventName == 'PLAYERREAGENTBANKSLOTS_CHANGED' then
-      updateBank = true
     elseif event.eventName == 'BAG_SORT' then
       if not InCombatLockdown() then
         sortBackpack = true
@@ -73,6 +69,8 @@ function refresh:StartUpdate()
       if not InCombatLockdown() then
         sortBackpackClassic = true
       end
+    elseif event.eventName == 'BAG_UPDATE_BANK' then
+      updateBank = true
     elseif const.BANK_BAGS[event.args[1]] then
       updateBank = true
     elseif const.REAGENTBANK_BAGS[event.args[1]] then
@@ -114,7 +112,7 @@ function refresh:StartUpdate()
   if updateBackpack then
     -- This timer runs during loading screens, which can cause the context
     -- to be cancelled before the draw even happens.
-    ctx:Timeout(30, function()
+    ctx:Timeout(60, function()
       self.isUpdateRunning = false
       items._preSort = false
     end)
@@ -145,11 +143,21 @@ function refresh:OnEnable()
   end)
 
   -- Register when bank slots change for any reason.
-  events:RegisterEvent('PLAYERBANKSLOTS_CHANGED', function()
+  events:RegisterEvent('PLAYERBANKSLOTS_CHANGED', function(_, slot)
     local ctx = context:New()
-    ctx:Set("wipe", false)
-    table.insert(refresh.UpdateQueue, {eventName = 'PLAYERBANKSLOTS_CHANGED', args = {}, ctx = ctx})
-    self:StartUpdate()
+    if slot > NUM_BANKGENERIC_SLOTS then
+      ctx:Set("wipe", true)
+    else
+      ctx:Set("wipe", false)
+    end
+    table.insert(refresh.UpdateQueue, {eventName = 'BAG_UPDATE_BANK', args = {}, ctx = ctx})
+  end)
+
+  -- Register when the bag slots change for any reason.
+  events:RegisterEvent('BAG_CONTAINER_UPDATE', function()
+    local ctx = context:New()
+    ctx:Set("wipe", true)
+    table.insert(refresh.UpdateQueue, {eventName = 'BAG_UPDATE', args = {}, ctx = ctx})
   end)
 
   -- Register when equipment sets change.
@@ -165,8 +173,7 @@ function refresh:OnEnable()
     events:RegisterEvent('PLAYERREAGENTBANKSLOTS_CHANGED', function()
       local ctx = context:New()
       ctx:Set("wipe", false)
-      table.insert(refresh.UpdateQueue, {eventName = 'PLAYERREAGENTBANKSLOTS_CHANGED', args = {}, ctx = ctx})
-      self:StartUpdate()
+      table.insert(refresh.UpdateQueue, {eventName = 'BAG_UPDATE_BANK', args = {}, ctx = ctx})
     end)
   end