+++ AddOns/BetterBags/CHANGELOG.md
@@ -1,6 +1,8 @@
 # BetterBags
 
-## [v0.1.61](https://github.com/Cidan/BetterBags/tree/v0.1.61) (2024-06-06)
-[Full Changelog](https://github.com/Cidan/BetterBags/compare/v0.1.60...v0.1.61) [Previous Releases](https://github.com/Cidan/BetterBags/releases)
+## [v0.1.75](https://github.com/Cidan/BetterBags/tree/v0.1.75) (2024-07-04)
+[Full Changelog](https://github.com/Cidan/BetterBags/compare/v0.1.74...v0.1.75) [Previous Releases](https://github.com/Cidan/BetterBags/releases)
 
-- Adjusted the inset for item frames slightly.  
+- Removed eyeball icon from category config.  
+    Right click menu is now used to hide or show categories.  
+    Dynamic categories always have the enabled backdrop.